backend/ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
deepface==0.0.79
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.26.2
python-multipart==0.0.6
tensorflow==2.15.0